+++ R/missingTable.R
@@ -12,7 +12,7 @@ missingTable <- function(obj,...)
     cg <- attr(obj, "x")[[1]]
   else 
     cg <- obj
-
+  
   varnames.orig <- attr(cg, "varnames.orig")
   Xext <- attr(cg, "Xext")
   temp <- Xext[varnames.orig]
@@ -27,15 +27,28 @@ missingTable <- function(obj,...)
 
   X<-attr(cg,"call")$call$formula
   X.eval <- eval(X)
-  if (inherits(X.eval, "formula"))
-    cg.update <- update(cg, data = Xext, simplify = FALSE)
-  else
-    cg.update <- update(cg, formula = Xext, simplify = FALSE)
   
-  if (inherits(obj, "createTable"))
-    ans <- update(obj, x = cg.update, hide.no = 'Avail',...)
-  else 
-    ans <- createTable(x = cg.update, hide.no = 'Avail',...)
+  if (inherits(X.eval, "formula"))
+    cg.update <- update(cg, data = Xext, simplify = FALSE,include.miss=FALSE)
+  else
+    cg.update <- update(cg, formula = Xext, simplify = FALSE,include.miss=FALSE)    
+
+  print(cg.update)
+  if (inherits(obj, "descrTable")){
+    ans <- createTable(cg.update, hide.no = 'Avail',...)
+    # if (inherits(X.eval, "formula"))
+    #   ans <- update(obj, data=Xext, hide.no = 'Avail', simplify = FALSE, include.miss=FALSE,...)
+    # else
+    #   ans <- update(obj, formula=Xext, hide.no = 'Avail', simplify = FALSE, include.miss=FALSE,...)  
+  }else {
+    
+    if (inherits(obj, "createTable")){
+      ans <- update(obj, x = cg.update, hide.no = 'Avail',...)  
+    } else{
+      ans <- createTable(x = cg.update, hide.no = 'Avail',...)  
+    }
+  }
+    
 
   class(ans)<-c("missingTable",class(ans))
   ans